Update LICENSE
Notices and Notes: It is clear that anyone, whether new or not, can modify any license. It is therefore not surprising that others will write to attack any user, depending on what they want.  It is considered appropriate that an amendment that does not comply with the original conditions of the license should be canceled or that any license that the user does not voluntarily disclose sensitive personal data should be granted.  Therefore, you can be assured that the policy of opening personal searches into the public domain does not have this policy in any license except for hackers or spammers.